+++ src/components/signup/confirm-payment-slide.tsx
