+++ src/components/signup/record-recovery-phrase-slide.tsx
