+++ src/components/signup/record-storage-handle-slide.tsx
